refactor for events, signUp
